+++ src/ServiceBus/ServiceBus/ChangeLog.md
@@ -19,7 +19,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
-
+* Fixed that `New-AzServiceBusAuthorizationRuleSASToken` returns invalid token. [#12975]
+ 
 ## Version 1.8.0
 * Added identity and encryption properties to New-AzServiceBusNamespace and Set-AzServiceBusNamespace.
 * Added New-AzServiceBusEncryptionConfig